2.2 work with v7.7.3

+++ src/storage/storage.cc
@@ -587,7 +587,7 @@ uint64_t Storage::GetTotalSize(const std::string &ns) {
   Redis::Database db(this, ns);
   uint64_t size, total_size = 0;
   uint8_t include_both =
-      rocksdb::DB::SizeApproximationFlags::INCLUDE_FILES | rocksdb::DB::SizeApproximationFlags::INCLUDE_MEMTABLES;
+      static_cast<uint8_t>(rocksdb::DB::SizeApproximationFlags::INCLUDE_FILES | rocksdb::DB::SizeApproximationFlags::INCLUDE_MEMTABLES);
   for (auto cf_handle : cf_handles_) {
     if (cf_handle == GetCFHandle(kPubSubColumnFamilyName) || cf_handle == GetCFHandle(kPropagateColumnFamilyName)) {
       continue;
